ARM: usb: chipidea: Leave FSL unmodified drivers.
We had previously merged these driver from linux-next. It's easier if we
follow the FSL release.

+++ drivers/usb/chipidea/bits.h
@@ -60,7 +60,7 @@
 #define DEVLC_PSPD_HS         (0x02UL << 25)
 #define DEVLC_PTW             BIT(27)
 #define DEVLC_STS             BIT(28)
-#define DEVLC_PTS(d)          (u32)(((d) & 0x7) << 29)
+#define DEVLC_PTS(d)          (((d) & 0x7) << 29)
 
 /* Encoding for DEVLC_PTS and PORTSC_PTS */
 #define PTS_UTMI              0
